App.tsx の責務分割・API URL定数化・URL二重パース解消
- App.tsx からロジックを useCircleExtraction カスタムフックに分離
- filterCircleLink をフック外の純粋関数に移動
- background/index.ts の API URL を定数 API_BASE_URL に切り出し
- content/index.tsx の filter+map を flatMap に統合し new URL() を1回に削減
